# Optimize: Use ImprovedDilatedAttention for single-GPU ring attention